Changes have been made to all files.
All HTML File changes:
We have added preconnect links and a stylesheet
link to the head of the documents for the
'Space-Grotesk' font. We have also added a
stylesheet link to the head of the documents for
'styles.css'.

All CSS File changes:
Styling has been added for all elements, mainly
involving the font using the * selector. We have
also added styling for the buttons and the
main-box div element.